Divide the create/edit card modal into its own component so that only one code base has to be maintained for this feature
