+++ frontend/src/pages/ContentPage/ConstructCardModal.css
